feat(metal): delta-net state buffers, and the f32-weight problem that dissolved (Phase 2)
The plan called for adding f32 weight access -- WeightSource exposes only fp16()/quant(), while
the delta-net kernel took norm_weight and A_log as const float*, as CUDA does. The intended fix
was a Metal-private f32 map mirroring CUDA's devf_.

Checking the converter first made the whole item disappear: convert_hf_to_bins casts EVERY tensor
to float16, so the container has no f32 to read. A private f32 map would have upcast fp16 to f32
and called it precision. The kernel now takes half for both, which is what the data actually is
and what every other weight on this path already is.

That does diverge from CUDA, which reads both as f32 straight from safetensors. a_log feeds
exp(-exp(a_log) * softplus(..)), so fp16 rounding there shows up as a small constant factor on a
layer's decay -- noted at the kernel as the first thing to suspect if a layer diverges that way.

State buffers: one contiguous fp32 allocation each for the conv window
([layers][conv_dim * (kernel-1)]) and the recurrent matrix
([layers][num_v_heads * key_hd * value_hd]), sliced by layer at dispatch. Zeroed at open --
the recurrence starts empty, and starting it from stale memory yields fluent nonsense, not a
crash. Both are zero-sized for models without linear-attention geometry.

LinearConv1d and LinearAttentionStep are now dispatched rather than refused; the vision four
still refuse, since they implement Gemma 4's tower and Qwen3.5's is a different architecture.

Gate is 10 checks 0 failed, metal_smoke 52 checks including the two-step state tests (which still
pass with the fp16 operands), and an existing model is unaffected at 541 tok in 161 ms.

+++ src/kernels/metal/70_linear_attn.metal
@@ -170,8 +170,14 @@ kernel void cpi_linear_attention_step(device const half* q [[buffer(0)]],
                                       device const half* z [[buffer(3)]],
                                       device const half* a [[buffer(4)]],
                                       device const half* b [[buffer(5)]],
-                                      device const float* norm_weight [[buffer(6)]],
-                                      device const float* a_log [[buffer(7)]],
+                                      // half, not float: the .ll2c container casts every weight
+                                      // to fp16, so there is no f32 source to read these from.
+                                      // CUDA reads them as f32 straight from safetensors, so the
+                                      // two backends differ here by fp16 rounding on a_log -- it
+                                      // feeds exp(-exp(a_log)*softplus(..)), so watch this first
+                                      // if a layer diverges by a small constant factor.
+                                      device const half* norm_weight [[buffer(6)]],
+                                      device const half* a_log [[buffer(7)]],
                                       device const half* dt_bias [[buffer(8)]],
                                       device float* recurrent_state [[buffer(9)]],
                                       device half* out [[buffer(10)]],
@@ -217,7 +223,7 @@ kernel void cpi_linear_attention_step(device const half* q [[buffer(0)]],
         shared_scalar[1] = rsqrt(tot + 1.0e-6f);
         shared_scalar[2] = lin_sigmoid(float(b[head]));
         shared_scalar[3] =
-            exp(-exp(a_log[head]) * lin_softplus(float(a[head]) + float(dt_bias[head])));
+            exp(-exp(float(a_log[head])) * lin_softplus(float(a[head]) + float(dt_bias[head])));
       }
     }
     threadgroup_barrier(mem_flags::mem_threadgroup);
@@ -273,7 +279,7 @@ kernel void cpi_linear_attention_step(device const half* q [[buffer(0)]],
   const float out_inv = shared_scalar[4];
 
   for (uint dv = lid; dv < vd_n; dv += nthr) {
-    out[v_base + dv] =
-        half(scratch[dv] * out_inv * norm_weight[dv] * lin_silu(float(z[v_base + dv])));
+    out[v_base + dv] = half(scratch[dv] * out_inv * float(norm_weight[dv]) *
+                            lin_silu(float(z[v_base + dv])));
   }
 }
